[NEON] Removed vreinterpret instructions
I saw it in the C intrinsics list, and thought that I really needed to
use it. Later, I thought, "why is that even needed...at all?", so I
looked it up, and it turnes out that the vreinterpret intrinsic is
actually a NOP. So *yay*.

Also, the compiler was like, "no, I'm not gonna compile these," so I had
to take them out. Also, they masked a int16->uint8 conversion problem
(it was doing uint16->uint8, so it didn't bound it correctly).

The big difference is that `vqmovn.u8` will convert an unsigned int16 to
an unsigned int 8, 'saturating'ly. However, if the initial value is <0,
then and it's treated like an unsigned int16, then it will wrap around,
be large, and result in '255', not '0'. `vqmovun.s16` does what I want
it to do.