Optimized code with modern data structures
Vector is slower than a LinkedList/ArrayList because synchronization (no
matter how fast it is, it has a non-zero overhead).
ArrayLists are usually better because of their lower memory overhead,
when compared to LinkedLists.
Collections.unmodifiableList is useful for returning coppies of lists.
HashTable is worse than HashMap for the same reason of unnecessary
synchronization.