Optimized conversion methods
Pointer access-and-increments are faster than array accesses, and more
easily optimized by the compiler.
Also changed how some of the math is done, hopefully streamlining the
process.
Even though the compiler should be able to do a few of the things that I
did, the generated binary shrunk, so I assume that it worked.
Also added compile-time flags to control the precision vs. speed of the
conversions, from bit shifts, to integer approximations, to full 32-bit
float stuff.
Also fixed some of the averaging formulae.