Simplified Y10B conversions
Now with 1 fewer copy!
Basically, I interleaved the unpack 10-bit => 16-bit routine with the
round 16-bit => 8-bit and store part, removing the need for an
intermediate buffer. It also works out that a few of the operations can
be combined for less total work.
Result: MUCH smaller memory overhead, and should be quite a bit faster
(like 2x).