Consolidate RGB<=>YUV conversion routines
A lot of duplicate code across many files.
Unfortunately, it's not apperent whether it's possible to refactor some
more and move it all to a few methods (mostly because it needs to be
fast).
However, all the math can be hidden behind a few macros, which is good.
Also threw in a few (very) minor fixes.