Merge OMX doAllocateBuffer/doUseBuffer
We get rid of a lookup per handler call (this is the part that has to be
fast) by storing the OMXFrameBuffer jobject in the OMX_FRAMEBUFFERTYPE's
pAppPrivate.
Also, we now have less complicated code, because the JNI methods
doAllocateBuffer & doUserBuffer were doing about 90% the same thing. And
the rest is now in an if statement.