Added test for double frame recycle
Should throw an exception.
It *is* possible for the frame to be un-recycled and reused while
preserving stored references, so checking if recycle() throws an
exception isn't a good measure of the frame's validity, but we should be
able to catch it at least some of the time.
This test is correct, because frames won't be un-recycled during the
callback (atm).