Realized how broken some of the accessor methods are
If called on a child, they just read/write to native memory; for
example, if we call write() on a child field accessor, it does NOT, in
fact, set a parameter/configuration on the component. It just sets a
field in native memory, which isn't what we want to do, because that
action is intentionally not visible from a client.